techpack: display: Make dimlayer transparent in LP1/LP2 power mode
To account for cases where dimlayer_hbm can be enabled while entering
AoD (and thus disabling Hight Brightness Mode), OPlus added code to make
the dimlayer fully transparent if the panel was in LP1 or LP2 mode and
the AoD layer is present. However, this is specific to OxygenOS; on
AOSP, we don't have that AoD layer, so this doesn't make the dimlayer
transparent for us.

Instead, let's remove all the AoD specific code and make it so that
any time the panel is in LP1 or LP2 mode, we make the dimlayer
transparent.

Change-Id: If396865ed38abd720ee8f2ed638d7b35ea9ef55f